Long live PhantomJS 2!

+++ src/breakpad/src/common/md5.h
@@ -18,7 +18,7 @@ struct MD5Context {
 
 void MD5Init(struct MD5Context *ctx);
 
-void MD5Update(struct MD5Context *ctx, unsigned char const *buf, unsigned len);
+void MD5Update(struct MD5Context *ctx, unsigned char const *buf, size_t len);
 
 void MD5Final(unsigned char digest[16], struct MD5Context *ctx);
