+++ src/breakpad/src/common/memory_range_unittest.cc
@@ -37,9 +37,9 @@ using testing::Message;
 
 namespace {
 
-const u_int32_t kBuffer[10] = { 0 };
+const uint32_t kBuffer[10] = { 0 };
 const size_t kBufferSize = sizeof(kBuffer);
-const u_int8_t* kBufferPointer = reinterpret_cast<const u_int8_t*>(kBuffer);
+const uint8_t* kBufferPointer = reinterpret_cast<const uint8_t*>(kBuffer);
 
 // Test vectors for verifying Covers, GetData, and Subrange.
 const struct {
@@ -55,13 +55,13 @@ const struct {
   { true, 2, kBufferSize - 2 },
   { true, kBufferSize - 1, 1 },
   { false, kBufferSize, 0 },
-  { false, kBufferSize, -1 },
+  { false, kBufferSize, static_cast<size_t>(-1) },
   { false, kBufferSize + 1, 0 },
-  { false, -1, 2 },
+  { false, static_cast<size_t>(-1), 2 },
   { false, 1, kBufferSize },
   { false, kBufferSize - 1, 2 },
-  { false, 0, -1 },
-  { false, 1, -1 },
+  { false, 0, static_cast<size_t>(-1) },
+  { false, 1, static_cast<size_t>(-1) },
 };
 const size_t kNumSubranges = sizeof(kSubranges) / sizeof(kSubranges[0]);
 
@@ -94,7 +94,7 @@ const size_t kNumElements = sizeof(kElements) / sizeof(kElements[0]);
 TEST(MemoryRangeTest, DefaultConstructor) {
   MemoryRange range;
   EXPECT_EQ(NULL, range.data());
-  EXPECT_EQ(0, range.length());
+  EXPECT_EQ(0U, range.length());
 }
 
 TEST(MemoryRangeTest, ConstructorWithDataAndLength) {
@@ -107,7 +107,7 @@ TEST(MemoryRangeTest, Reset) {
   MemoryRange range;
   range.Reset();
   EXPECT_EQ(NULL, range.data());
-  EXPECT_EQ(0, range.length());
+  EXPECT_EQ(0U, range.length());
 
   range.Set(kBuffer, kBufferSize);
   EXPECT_EQ(kBufferPointer, range.data());
@@ -115,7 +115,7 @@ TEST(MemoryRangeTest, Reset) {
 
   range.Reset();
   EXPECT_EQ(NULL, range.data());
-  EXPECT_EQ(0, range.length());
+  EXPECT_EQ(0U, range.length());
 }
 
 TEST(MemoryRangeTest, Set) {
@@ -126,14 +126,14 @@ TEST(MemoryRangeTest, Set) {
 
   range.Set(NULL, 0);
   EXPECT_EQ(NULL, range.data());
-  EXPECT_EQ(0, range.length());
+  EXPECT_EQ(0U, range.length());
 }
 
 TEST(MemoryRangeTest, SubrangeOfEmptyMemoryRange) {
   MemoryRange range;
   MemoryRange subrange = range.Subrange(0, 10);
   EXPECT_EQ(NULL, subrange.data());
-  EXPECT_EQ(0, subrange.length());
+  EXPECT_EQ(0U, subrange.length());
 }
 
 TEST(MemoryRangeTest, SubrangeAndGetData) {
@@ -156,7 +156,7 @@ TEST(MemoryRangeTest, SubrangeAndGetData) {
       EXPECT_FALSE(range.Covers(sub_offset, sub_length));
       EXPECT_EQ(NULL, range.GetData(sub_offset, sub_length));
       EXPECT_EQ(NULL, subrange.data());
-      EXPECT_EQ(0, subrange.length());
+      EXPECT_EQ(0U, subrange.length());
     }
   }
 }
