+++ src/breakpad/src/tools/linux/md2core/minidump_memory_range_unittest.cc
@@ -38,9 +38,9 @@ using testing::Message;
 
 namespace {
 
-const u_int32_t kBuffer[10] = { 0 };
+const uint32_t kBuffer[10] = { 0 };
 const size_t kBufferSize = sizeof(kBuffer);
-const u_int8_t* kBufferPointer = reinterpret_cast<const u_int8_t*>(kBuffer);
+const uint8_t* kBufferPointer = reinterpret_cast<const uint8_t*>(kBuffer);
 
 // Test vectors for verifying Covers, GetData, and Subrange.
 const struct {
@@ -56,13 +56,13 @@ const struct {
   { true, 2, kBufferSize - 2 },
   { true, kBufferSize - 1, 1 },
   { false, kBufferSize, 0 },
-  { false, kBufferSize, -1 },
+  { false, kBufferSize, static_cast<size_t>(-1) },
   { false, kBufferSize + 1, 0 },
-  { false, -1, 2 },
+  { false, static_cast<size_t>(-1), 2 },
   { false, 1, kBufferSize },
   { false, kBufferSize - 1, 2 },
-  { false, 0, -1 },
-  { false, 1, -1 },
+  { false, 0, static_cast<size_t>(-1) },
+  { false, 1, static_cast<size_t>(-1) },
 };
 const size_t kNumSubranges = sizeof(kSubranges) / sizeof(kSubranges[0]);
 
@@ -95,7 +95,7 @@ const size_t kNumElements = sizeof(kElements) / sizeof(kElements[0]);
 TEST(MinidumpMemoryRangeTest, DefaultConstructor) {
   MinidumpMemoryRange range;
   EXPECT_EQ(NULL, range.data());
-  EXPECT_EQ(0, range.length());
+  EXPECT_EQ(0U, range.length());
 }
 
 TEST(MinidumpMemoryRangeTest, ConstructorWithDataAndLength) {
@@ -108,7 +108,7 @@ TEST(MinidumpMemoryRangeTest, Reset) {
   MinidumpMemoryRange range;
   range.Reset();
   EXPECT_EQ(NULL, range.data());
-  EXPECT_EQ(0, range.length());
+  EXPECT_EQ(0U, range.length());
 
   range.Set(kBuffer, kBufferSize);
   EXPECT_EQ(kBufferPointer, range.data());
@@ -116,7 +116,7 @@ TEST(MinidumpMemoryRangeTest, Reset) {
 
   range.Reset();
   EXPECT_EQ(NULL, range.data());
-  EXPECT_EQ(0, range.length());
+  EXPECT_EQ(0U, range.length());
 }
 
 TEST(MinidumpMemoryRangeTest, Set) {
@@ -127,14 +127,14 @@ TEST(MinidumpMemoryRangeTest, Set) {
 
   range.Set(NULL, 0);
   EXPECT_EQ(NULL, range.data());
-  EXPECT_EQ(0, range.length());
+  EXPECT_EQ(0U, range.length());
 }
 
 TEST(MinidumpMemoryRangeTest, SubrangeOfEmptyMemoryRange) {
   MinidumpMemoryRange range;
   MinidumpMemoryRange subrange = range.Subrange(0, 10);
   EXPECT_EQ(NULL, subrange.data());
-  EXPECT_EQ(0, subrange.length());
+  EXPECT_EQ(0U, subrange.length());
 }
 
 TEST(MinidumpMemoryRangeTest, SubrangeAndGetData) {
@@ -157,7 +157,7 @@ TEST(MinidumpMemoryRangeTest, SubrangeAndGetData) {
       EXPECT_FALSE(range.Covers(sub_offset, sub_length));
       EXPECT_EQ(NULL, range.GetData(sub_offset, sub_length));
       EXPECT_EQ(NULL, subrange.data());
-      EXPECT_EQ(0, subrange.length());
+      EXPECT_EQ(0U, subrange.length());
     }
   }
 }
@@ -185,7 +185,7 @@ TEST(MinidumpMemoryRangeTest, SubrangeWithMDLocationDescriptor) {
       EXPECT_FALSE(range.Covers(sub_offset, sub_length));
       EXPECT_EQ(NULL, range.GetData(sub_offset, sub_length));
       EXPECT_EQ(NULL, subrange.data());
-      EXPECT_EQ(0, subrange.length());
+      EXPECT_EQ(0U, subrange.length());
     }
   }
 }
@@ -222,7 +222,7 @@ TEST(MinidumpMemoryRangeTest, GetArrayElmentWithTemplateType) {
 }
 
 TEST(MinidumpMemoryRangeTest, GetAsciiMDString) {
-  u_int8_t buffer[100] = { 0 };
+  uint8_t buffer[100] = { 0 };
 
   MDString* md_str = reinterpret_cast<MDString*>(buffer);
   md_str->length = 4;
@@ -233,7 +233,7 @@ TEST(MinidumpMemoryRangeTest, GetAsciiMDString) {
   md_str->buffer[4] = '\0';
 
   size_t str2_offset =
-      sizeof(MDString) + (md_str->length + 1) * sizeof(u_int16_t);
+      sizeof(MDString) + (md_str->length + 1) * sizeof(uint16_t);
 
   md_str = reinterpret_cast<MDString*>(buffer + str2_offset);
   md_str->length = 9;  // Test length larger than actual string
